Agregada la funcionalidad para crear y unirse a grupos, añadir amigos y mostrar las ultimas actividades de los amigos.

+++ ApiServer/DTOs/Responses/FriendsFrontPage.cs
@@ -11,6 +11,6 @@ namespace ApiServer.DTOs.Responses
         public string Type { get; set; }
         public DateTime Start { get; set; }
         public string Route { get; set; }
-        public float Distance { get; set; }
+        public double Distance { get; set; }
     }
 }
